Moved crossref utils to separate module
